refactor: improve admin route security
- Create centralized adminAuth utility for secure admin verification

- Remove service role key from middleware

- Implement two-layer admin verification (middleware + API)

- Update all admin routes to use secure verification

- Add proper TypeScript types
