+++ src/app/api/admin/check-access/route.ts
@@ -1,40 +1,15 @@
-import { createRouteHandlerClient } from '@supabase/auth-helpers-nextjs';
-import { cookies } from 'next/headers';
 import { NextResponse } from 'next/server';
+import { verifyAdminAccess } from '@/utils/adminAuth';
 
 export async function GET() {
   try {
-    const cookieStore = cookies();
-    const supabase = createRouteHandlerClient(
-      { cookies: () => cookieStore },
-      {
-        supabaseKey: process.env.SUPABASE_SERVICE_ROLE_KEY,
-      }
-    );
+    const result = await verifyAdminAccess();
 
-    const {
-      data: { session },
-    } = await supabase.auth.getSession();
-
-    if (!session) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    if ('error' in result) {
+      return NextResponse.json({ error: result.error }, { status: result.status });
     }
 
-    // Get user with service role to ensure we have the latest metadata
-    const {
-      data: { user },
-      error: userError,
-    } = await supabase.auth.admin.getUserById(session.user.id);
-
-    if (userError || !user) {
-      return NextResponse.json({ error: 'Failed to verify access' }, { status: 500 });
-    }
-
-    const isAdmin = user.user_metadata?.is_admin === true;
-
-    if (!isAdmin) {
-      return NextResponse.json({ error: 'Access denied' }, { status: 403 });
-    }
+    const { user } = result;
 
     // Return user data for client-side use
     return NextResponse.json({
