+++ src/app/api/admin/users/route.ts
@@ -1,25 +1,15 @@
-import { createRouteHandlerClient } from '@supabase/auth-helpers-nextjs';
-import { cookies } from 'next/headers';
 import { NextResponse } from 'next/server';
+import { verifyAdminAccess } from '@/utils/adminAuth';
 
 export async function GET() {
   try {
-    const cookieStore = cookies();
-    const supabase = createRouteHandlerClient(
-      { cookies: () => cookieStore },
-      {
-        supabaseKey: process.env.SUPABASE_SERVICE_ROLE_KEY,
-      }
-    );
+    const result = await verifyAdminAccess();
 
-    // First check if the requesting user is an admin
-    const {
-      data: { session },
-    } = await supabase.auth.getSession();
-
-    if (!session?.user?.user_metadata?.is_admin) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    if ('error' in result) {
+      return NextResponse.json({ error: result.error }, { status: result.status });
     }
+
+    const { supabase } = result;
 
     // Get all users
     const {
